requirements.txt
# NexusWAF RL Model — Python dependencies
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# RL Environment
gymnasium>=0.29.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0   # SMOTE for class balancing

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Visualization & evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP log parsing
urllib3>=2.0.0

# Config & serialization
pyyaml>=6.0
